Extract execute function
